feat: add bounding box, letterbox, and drawing utilities
New features:
- Bounding Box Utils: convertBoxFormat, scaleBoxes, clipBoxes, calculateIoU, nonMaxSuppression
- Letterbox Utils: letterbox padding for YOLO-style models, reverseLetterbox for coordinate mapping
- Drawing Utils: drawBoxes, drawKeypoints, overlayMask, overlayHeatmap for visualizations

Implementation:
- Added TypeScript types and wrapper functions in src/index.tsx and src/types.ts
- Added native spec methods in src/NativeVisionUtils.ts
- iOS: BoundingBoxUtils.swift, LetterboxUtils.swift, DrawingUtils.swift
- Android: BoundingBoxUtilsAndroid.kt, LetterboxUtilsAndroid.kt, DrawingUtilsAndroid.kt
- Fixed Android module signatures to match TurboModule spec
- Fixed CameraFrameAndroid bitwise operation on doubles

+++ android/src/main/java/com/visionutils/CameraFrameAndroid.kt
@@ -85,7 +85,7 @@ object CameraFrameAndroid {
         var tensorData: DoubleArray = if (outputFormat == "grayscale") {
             rgbToGrayscale(rgbData, outputWidth, outputHeight)
         } else {
-            DoubleArray(rgbData.size) { rgbData[it].toDouble() and 0xFF.toDouble() }
+            DoubleArray(rgbData.size) { (rgbData[it].toInt() and 0xFF).toDouble() }
         }
 
         // Apply normalization
